+++ components/Game.py
@@ -1,89 +1,43 @@
 import pygame 
 from pygame import  *
 from components.Board import Board
-from components.Player import Player
-from components.Piece import Piece
-import math
+from components.HelperModule import HelperModule
 
 class Game:
 
-    def __init__(self) -> None:
-        # game var
-        self.id = "game_one"                        # str: uuid?
-        self.width = 800
-        self.height = 800
-        self.tile_size = self.width/8
-        self.board = Board(self.width, self.height, self.tile_size)                # object
-        self.pieces = self.populate_board()         # array[object*]: contenant les 32 pièces
-        #self.players = [Player(0), Player(1)]       # array[object*]: contenant les 2 joueurs
-        self.turn_count = 0                         # int: le nombre de tour
-        self.active_player = 0                      # int: player_id
-        self.active_piece = None                    # str: piece_id
+    def __init__(self, game_id) -> None:
+        # app var
+        self.id = game_id                                                                       # str: uuid?
+        self.display_width = 1200                                                               # int: px
+        self.display_height = 900                                                               # int: px
+        self.board_width = 800                                                                  # int: px
+        self.board_height = 800                                                                 # int: px
+        self.tile_size = self.board_width/8                                                     # int: px
+        self.board = Board("board_one", self.board_width, self.board_height, self.tile_size)    # Surface   
+        self.helper = HelperModule("helper")                                                    # Object: helper method
 
         # pygame var
         self.instance = pygame.init()
         self.title = pygame.display.set_caption("Py chess")
         self.clock = pygame.time.Clock()
-        self.display = pygame.display.set_mode((800,800))
+        self.display = pygame.display.set_mode((self.display_width,self.display_height))
 
+        # game var
+        self.turn_count = 0                         # int: le nombre de tour
+        self.active_player = 0                      # int: player_id
+        self.active_piece = None                    # str: piece_id
+
+        # starting game
         self.loop = self.main_loop()
-
-    # place les pièces sur le plateau de jeu
-    def populate_board(self):
-        pieces = [
-            # id, type, color, chess_coordinates, unit_size(px), index
-            # white pawns
-            Piece("white_pawn_1", 0, 0, "A2", self.tile_size, 0),
-            Piece("white_pawn_2", 0, 0, "B2", self.tile_size, 1),
-            Piece("white_pawn_3", 0, 0, "C2", self.tile_size, 2),
-            Piece("white_pawn_4", 0, 0, "D2", self.tile_size, 3),
-            Piece("white_pawn_5", 0, 0, "E2", self.tile_size, 4),
-            Piece("white_pawn_6", 0, 0, "F2", self.tile_size, 5),
-            Piece("white_pawn_7", 0, 0, "G2", self.tile_size, 6),
-            Piece("white_pawn_8", 0, 0, "H2", self.tile_size, 7),
-            # black pawns
-            Piece("black_pawn_1", 0, 1, "A7", self.tile_size, 8),
-            Piece("black_pawn_2", 0, 1, "B7", self.tile_size, 9),
-            Piece("black_pawn_3", 0, 1, "C7", self.tile_size, 10),
-            Piece("black_pawn_4", 0, 1, "D7", self.tile_size, 11),
-            Piece("black_pawn_5", 0, 1, "E7", self.tile_size, 12),
-            Piece("black_pawn_6", 0, 1, "F7", self.tile_size, 13),
-            Piece("black_pawn_7", 0, 1, "G7", self.tile_size, 14),
-            Piece("black_pawn_8", 0, 1, "H7", self.tile_size, 15),
-            # white knight
-            Piece("white_knight_1", 1, 0, "B1", self.tile_size, 16),
-            Piece("white_knight_2", 1, 0, "G1", self.tile_size, 17),
-            # black knight
-            Piece("black_knight_1", 1, 1, "B8", self.tile_size, 18),
-            Piece("black_knight_2", 1, 1, "G8", self.tile_size, 19),
-            # white bishop
-            Piece("white_bishop_1", 2, 0, "C1", self.tile_size, 20),
-            Piece("white_bishop_2", 2, 0, "F1", self.tile_size, 21),
-            # black bishop
-            Piece("black_bishop_1", 2, 1, "C8", self.tile_size, 22),
-            Piece("black_bishop_2", 2, 1, "F8", self.tile_size, 23),
-            # white rook
-            Piece("white_rook_1", 3, 0, "A1", self.tile_size, 24),
-            Piece("white_rook_1", 3, 0, "H1", self.tile_size, 25),
-            # black rook
-            Piece("black_rook_1", 3, 1, "A8", self.tile_size, 26),
-            Piece("black_rook_2", 3, 1, "H8", self.tile_size, 27),
-            # white queen
-            Piece("white_queen_1", 4, 0, "D1", self.tile_size, 28),
-            # black queen
-            Piece("black_queen_1", 4, 1, "D8", self.tile_size, 29),
-            # white king
-            Piece("white_king_1", 5, 0, "E1", self.tile_size, 30),
-            # black king
-            Piece("black_king_1", 5, 1, "E8", self.tile_size, 31),
-        ]
-        return pieces
 
     # la loop logique du jeu
     def main_loop(self):
         while True:
             # player inputs
             self.inputs()
+
+            # do thing
+            self.logic()
 
             # render
             self.render()
@@ -146,29 +100,9 @@ class Game:
                         temp_piece.selected = False
                         self.active_piece = None
 
-    # Piece object: récupère la pièce grâce à son id
-    def get_piece_by_id(self, piece_id):
-        for piece in self.pieces:
-            if piece_id != None and piece.id == piece_id:
-                return piece
-            elif piece_id == None and piece.id == self.active_piece:
-                return piece
- 
-        print("no Piece found")
-
-    # String: récupère les coordonnées de la case du jeu d'échec "A1" "B1" etc
-    def get_chess_coordinates(self, coordinates):
-        xy_coordinates = self.get_xy_coordinates(coordinates)
-        for key, value in self.board.tiles_xy.items():
-            if value == xy_coordinates:
-                return key
-        return "key not found"
-
-    # tupple( x:int, y:int): récupère les xy coordonnées de la case 
-    def get_xy_coordinates(self, coordinates):
-        x = math.floor(coordinates[0]/self.tile_size)
-        y = math.floor(coordinates[1]/self.tile_size)
-        return (x, y)
+    # logic
+    def logic(self):
+        pass
 
     # dessine le jeu
     def render(self):
@@ -178,7 +112,8 @@ class Game:
 
         # dessine le plateau et les pièces
         self.board.draw(self.display)
-        for piece in self.pieces:
+
+        for piece in self.board.pieces_list:
             piece.draw(self.display)
 
             # si une pièce est selectionnée, place un curseur jaune sur la case
