+++ main.py
@@ -3,6 +3,6 @@ from components.Game import Game
 class App:
 
     def __init__(self) -> None:
-        self.game = Game()
+        self.game = Game("game_one")
         
 app = App()
